#79 update : 그래프 저장 공유 로직 util로 빼기
